requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
sqlmodel>=0.0.16
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
alembic>=1.13.1
pytest>=8.3.3
pytest-asyncio>=0.23.5
asyncpg>=0.29.0